// src/enums/DynamicField.ts
export enum DynamicField {
    Brand = 0,
    Organization = 1,
    OrganizationAddress = 2,
    OrganizationTel = 3,
    OrganizationFax = 4,
    OrganizationWebsite = 5,
    Title = 6,
    PatientInfo = 7,
    Fullname = 8,
    YearOfBirth = 9,
    Gender = 10,
    Address = 11,
    Diagnostic = 12,
    // 13
    Result = 14,
    // 15, 16, 17, 18
    Conclude = 19,
    Recommend = 20,
    Datetime = 21,
    Doctor = 22,
    DoctorName = 23,
    DoctorDD = 24,
}